Extract token authorization function
